[AIRFLOW-5253] Move GCP KMS to core (#5859)
This commit moves GCP KMS from contrib to core.
For more information check AIP-21.
